+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # PyLauncher initial commit
 
+# 07/04/2024
+- Profiles view works
+- Profiles edit working
+- Added add_profile_dict to config.py
+- Added optional save argument to remove_profile in config.py
+- While i saw how Titan Launcher works, i noticed that uses allowedReleaseTypes. So i will modify config.py
+
+# 06/04/2024
+- Added format_icon, parse_icon to config.py
+- Fixed typo: get_profile_names -> get_profiles_names on config.py
+- New file on src/modules/widgets/profiles
+- Working on profiles page
+
+# Idk when i updted this
 First commit, almost nothin work, but the ui look well.
 
 What's working?: